Give user feedback on settings page

+++ src/components/settings/DeleteDataButton.vue
@@ -1,17 +1,28 @@
 <template>
   <div>
-    <button @click="clearStorage">Clear localStorage</button>
+    <button @click="clearStorage">
+      Delete Walnut.Direct Local Storage
+    </button>
   </div>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
+import { showMessageOnNextPageReload } from "@/main";
 
 export default Vue.extend({
   methods: {
     clearStorage() {
-      localStorage.clear();
-      window.location.reload();
+      if (
+        // eslint-disable-next-line no-alert
+        window.confirm(
+          "Are you sure you want to clear local storage? This will reset your schedule and your settings."
+        )
+      ) {
+        localStorage.clear();
+        showMessageOnNextPageReload("Cleared local storage");
+        window.location.reload();
+      }
     },
   },
 });
